requirements.txt
python-telegram-bot==20.3
fpdf==1.7.2
uvicorn
fastapi